Fixing af_get_last_error for unified backend
The way it was implemented, it was getting the last error
generated from  unified backend. This change gets the last error
generated from a particular backend instead.
